Add files via upload
Tried another approach in the "try.js" file, by working with an array and afterwards assigning each value to the HTML collection, since one can't use all the wonderful Array methods on html collections.
Looks much more promising.
Still some bugs to fix: 
No Game Over yet
no sounds
fruit can randomly appear on the same square as the snake is moving on
Returns an error as soon as the snake moves outside the field on top or bottom, but doesn't when it's moving to the right or left. 
The snake can overlap its own body.